Switch default setting to not use jax and refactor integrate_diffrax method
The primary change in this commit is switching the default setting of 'use_jax' from True to False. This change affected various parts of the code, resulting in an overall function refactor of the 'integrate_diffrax' method and implementation of 'partial' and 'jit' or their equivalent operations according to the 'use_jax' flag. Additionally, an extra layer of functionality has been added to the 'integrate_diffrax' method when 'use_jax' is set to True.