Version currently working for JAX CPU (on the Mac) Refine parameters and improve methods in ukf_voss.py
The updates on ukf_voss.py include the refinement of the tolerance for the stepsize_controller in the evolve_f_jax method, giving it a more suitable precision. The script now also uses args instead of p in ode_func function and the evolve_f_jax function is now decorated for jax.jit. The Pxy initialization has also been changed to use jax, providing a more consistent usage of libraries in the code.  Additionally, scipy has been imported for debugging purposes.