Updated initial conditions and loop ranges in 'ukf_voss.py' and 'Assignment6.ipynb'
Trying to add a control loop. Not working yet
Made modifications to the initial conditions and loop ranges under functions in 'ukf_voss.py' file. Additionally, introduced new variables and altered image rendering in 'Assignment6.ipynb'. This ensures that the implementation adjusts to dynamic inputs and provides accuracy in the output.